Use new SuperList for SuperInventories
