+++ src/main/java/net/glowstone/util/collection/SuperIterator.java
@@ -1,7 +1,5 @@
 package net.glowstone.util.collection;
 
-import net.glowstone.util.Adapter;
-
 import java.util.Iterator;
 import java.util.NoSuchElementException;
 
@@ -9,32 +7,22 @@ import java.util.NoSuchElementException;
  * An Iterator which delegates in other iterators.
  */
 public class SuperIterator<E> implements Iterator<E> {
-    private final Iterator parentIterator;
+    private final Iterator<Iterable<E>> parentIterator;
     private Iterator<E> childIterator;
-    private Adapter adapter;
 
     public SuperIterator(Iterable<Iterable<E>> iterable) {
         this(iterable.iterator());
     }
 
     public SuperIterator(Iterator<Iterable<E>> parentIterator) {
-        this(parentIterator, IterableIteratorAdapter.INSTANCE);
-    }
-
-    public SuperIterator(Iterable iterable, Adapter<?, Iterator> adapter) {
-        this(iterable.iterator(), adapter);
-    }
-
-    public SuperIterator(Iterator parentIterator, Adapter<?, Iterator> adapter) {
         this.parentIterator = parentIterator;
-        this.adapter = adapter;
     }
 
     @Override
     public boolean hasNext() {
         while (childIterator == null || !childIterator.hasNext()) {
             if (parentIterator.hasNext()) {
-                childIterator = (Iterator<E>) adapter.adapt(parentIterator.next());
+                childIterator = parentIterator.next().iterator();
             } else {
                 return false;
             }
